requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.4.1
Flask-Cors==4.0.0
SpeechRecognition==3.10.0
pydub==0.25.1